使用xtra_rhel6.x代替自身的库 升级Thrift RPC到0.11.0接口 jsoncpp index的bug RabbitMQ封装添加默认的生产者、消费者设置函数

+++ source/client/TzMonitorHttpClientHelper.cpp
@@ -192,7 +192,9 @@ public:
             }
 
             std::vector<event_info_t> info;
-            for (size_t i = 0; i < infoList.size(); i++) {
+			
+			// jsoncpp的bug，只能使用int或者Json::Value::ArrayIndex，不能使用size_t作为索引
+            for (int i = 0; i < infoList.size(); i++) {  
                 if (!infoList[i]["count"].isString() || !infoList[i]["value_sum"].isString() ||
                     !infoList[i]["value_avg"].isString() || !infoList[i]["value_std"].isString() )
                 {
